% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gwas_example}
\alias{gwas_example}
\title{Example GWAS Data}
\format{
A data frame with 99226 rows and 4 variables:
\describe{
  \item{SNP}{SNP identifier}
  \item{CHR}{Chromosome number (1-25)}
  \item{BP}{Base pair position}
  \item{P}{P-value from association test}
}
}
\source{
Simulated data for package demonstration.
}
\usage{
gwas_example
}
\description{
A simulated dataset containing GWAS results for demonstration purposes.
}
\examples{
data(gwas_example)
head(gwas_example)
ggManhattan(gwas_example)
}
\keyword{datasets}
